Extract notes-tree-state region into sections/notes-tree-state.js
Sixth and final region extraction of P-08. Narrow cut — just the _NOTES_OPEN_KEY constant, restoreNotesTreeState, and the global toggle listener (now wrapped in initNotesTreeStateSideEffects). _NOTES_OPEN_KEY is re-exported because uploadToNotes + createNotesSubdir still seed localStorage by hand; moving those callers is P-09 territory.
